api/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
google-generativeai>=0.3.2
pydantic>=2.0.0
python-multipart>=0.0.6
matplotlib>=3.7.0
python-dotenv>=0.19.0
